64 bit conversion

+++ freex3d/src/lib/input/EAIHeaders.h
@@ -61,7 +61,7 @@ void EAI_RW(char *bufptr);
 void Parser_deleteParserForScanStringValueToMem(void);			/* from EAI_C_CommonFunctions.c */
 void Parser_scanStringValueToMem(struct X3D_Node *node, size_t coffset, indexT ctype, char *value, int isXML);
 									/* from EAI_C_CommonFunctions.c */
-int returnRoutingElementLength(int);					/* from EAI_C_CommonFunctions.c */
+size_t returnRoutingElementLength(int);					/* from EAI_C_CommonFunctions.c */
 void createLoadURL(char *);						/* from EAIEventsIn.c */
 void EAI_parse_commands(void);						/* from EAIEventsIn.c */
 void EAI_Anchor_Response(int);						/* from EAIEventsIn.c */
